Move proj_units_to_meters() to satpy.utils
